+++ apps/prairielearn/src/api/v1/index.ts
@@ -18,7 +18,7 @@ const authzHasCourseInstanceView = asyncHandler(async (req, res, next) => {
   next();
 });
 
-// Pretty-print all JSON responses
+// Pretty-print all JSON responses.
 router.use((await import('./prettyPrintJson.js')).default);
 
 // All course instance pages require authorization
@@ -57,13 +57,13 @@ router.use(
   (await import('./endpoints/courseInstanceAccessRules/index.js')).default,
 );
 
-// If no earlier routes matched, 404 the route
+// If no earlier routes matched, 404 the route.
 router.use((await import('./notFound.js')).default);
 
 // The Sentry error handler must come before our own.
 router.use(Sentry.expressErrorHandler());
 
-// Handle errors independently from the normal PL error handling
+// Handle errors independently from the normal PL error handling.
 router.use((await import('./error.js')).default);
 
 export default router;