+++ apps/prairielearn/src/batched-migrations/20241018171312_questions__share_publicly__backfill.sql
@@ -1,0 +1,8 @@
+-- BLOCK backfill_share_publicly
+UPDATE questions
+SET
+  share_publicly = shared_publicly
+WHERE
+  shared_publicly IS TRUE
+  AND id >= $start
+  AND id <= $end;