+++ apps/prairielearn/src/components/SubmissionPanel.html.ts
@@ -1,7 +1,7 @@
 import { differenceInMilliseconds } from 'date-fns';
 import { z } from 'zod';
 
-import { HtmlValue, html, unsafeHtml } from '@prairielearn/html';
+import { type HtmlValue, html, unsafeHtml } from '@prairielearn/html';
 
 import { config } from '../lib/config.js';
 import {
@@ -254,7 +254,7 @@ export function SubmissionPanel({
             : ''}"
           data-submission-id="${submission.id}"
           id="submission-${submission.id}-body"
-          ${submissionHtml === undefined && question.type === 'Freeform'
+          ${question.type === 'Freeform'
             ? html`
                 data-dynamic-render-url="${renderUrlPrefix}/variant/${variant_id}/submission/${submission.id}"
               `