+++ apps/prairielearn/src/ee/pages/instructorAiGenerateQuestion/instructorAiGenerateQuestion.html.ts
@@ -5,24 +5,34 @@ import { Navbar } from '../../../components/Navbar.html.js';
 
 const examplePrompts = [
   {
-    id: 'random_mc',
-    question:
-      'Write a multiple choice question asking the user to choose the median of 5 random numbers between 1 and 100. Display all of the numbers to the user, and then make each of the numbers a potential answer, with the median being the correct answer.',
+    id: 'Select median of 5 random numbers',
+    promptGeneral:
+      'Write a multiple choice question asking the user to choose the median of 5 random numbers between 1 and 100. Display all numbers to the user, and ask them to choose the median.',
+    promptUserInput:
+      'Each random number generated should be a potential answer to the multiple-choice question. Randomize the order of the numbers.',
+    promptGrading: 'The correct answer is the median of the numbers.',
   },
   {
-    id: 'random_int',
-    question:
-      'Write a question that asks the user to multiply two integers. You should randomly generate two integers A and B, display them to the user, and then ask the user to provide the product C = A * B in an integer input box. The correct answer should be the product of the two numbers that you generated.',
+    id: 'Sum random integers',
+    promptGeneral:
+      'Write a question that asks the user to multiply two integers. You should randomly generate two integers A and B, display them to the user, and then ask the user to provide the product C = A * B.',
+    promptUserInput: 'Provide an integer input box for the user to enter the product.',
+    promptGrading: 'The correct answer is the product of A and B.',
   },
   {
-    id: 'basic_int',
-    question:
-      'Write a question asking "What The Answer to the Ultimate Question of Life, the Universe, and Everything?". Provide an integer box for the user to answer. The correct answer is 42.',
+    id: 'Answer to Ultimate Question',
+    promptGeneral:
+      'Write a question asking "What Is The Answer to the Ultimate Question of Life, the Universe, and Everything?".',
+    promptUserInput: 'Provide an integer box for the user to answer.',
+    promptGrading: 'The correct answer is 42.',
   },
   {
-    id: 'implicit_specified_physics',
-    question:
+    id: 'Calculate Projectile Distance',
+    promptGeneral:
       'Write a question that asks the user to calculate how far a projectile will be launched. Display to the user an angle randomly generated between 30 and 60 degrees, and a velocity randomly generated between 10 and 20 m/s, and ask for the distance (in meters) that the object travels assuming no wind resistance.',
+    promptUserInput: 'Provide a numerical input box for the user to enter an answer.',
+    promptGrading:
+      'The correct answer is the distance that the projectile will travel, using the corresponding formula.',
   },
 ];
 
@@ -50,8 +60,48 @@ export function AiGeneratePage({ resLocals }: { resLocals: Record<string, any> }
                 <input type="hidden" name="__csrf_token" value="${resLocals.__csrf_token}" />
                 <input type="hidden" name="__action" value="generate_question" />
                 <div class="form-group">
-                  <label for="user-prompt-llm">Prompt:</label>
+                  <label for="user-prompt-llm">
+                    Please give a high-level overview of the question. What internal parameters need
+                    to be generated and what information do we provide to students?
+                  </label>
                   <textarea name="prompt" id="user-prompt-llm" class="form-control"></textarea>
+                  <div class="form-text form-muted">
+                    <em>
+                      Example: A toy car is pushed off a table with height h at speed v0. Assume
+                      acceleration due to gravity as 9.81 m/s^2. H is a number with 1 decimal digit
+                      selected at random between 1 and 2 meters. V0 is a an integer between 1 and 4
+                      m/s. How long does it take for the car to reach the ground?
+                    </em>
+                  </div>
+                </div>
+                <div class="form-group">
+                  <label for="user-prompt-llm-user-input">
+                    How should students input their solution? What choices or input boxes are they
+                    given?
+                  </label>
+                  <textarea
+                    name="prompt_user_input"
+                    id="user-prompt-llm-user-input"
+                    class="form-control"
+                  ></textarea>
+                  <div class="form-text form-muted">
+                    <em>
+                      Example: students should enter the solution using a decimal number. The answer
+                      should be in seconds.
+                    </em>
+                  </div>
+                </div>
+                <div class="form-group">
+                  <label for="user-prompt-llm-grading">
+                    How do we determine if a solution is correct/which answer choice is correct?
+                  </label>
+                  <textarea name="prompt_grading" id="user-prompt-llm-grading" class="form-control">
+                  </textarea>
+                  <div class="form-text form-muted">
+                    <em>
+                      Example: the answer is computed as sqrt(2 * h / g) where g = 9.81 m/s^2
+                    </em>
+                  </div>
                 </div>
                 <button class="btn btn-primary">
                   <span
@@ -59,15 +109,28 @@ export function AiGeneratePage({ resLocals }: { resLocals: Record<string, any> }
                     role="status"
                     aria-hidden="true"
                     data-loading-class-remove="d-none"
-                  ></span>
+                  >
+                  </span>
                   Create question
                 </button>
               </form>
               Or choose a question from our list of example prompts:
-              <select id="user-prompt-example" onchange="setPromptToExample()">
+              <select
+                id="user-prompt-example"
+                onchange="setPromptToExample()"
+                class="custom-select"
+              >
                 <option value=""></option>
                 ${examplePrompts.map(
-                  (question) => html`<option value="${question.question}">${question.id}</option>`,
+                  (question) =>
+                    html`<option
+                      value="${question.id}"
+                      data-prompt-general="${question.promptGeneral}"
+                      data-prompt-user-input="${question.promptUserInput}"
+                      data-prompt-grading="${question.promptGrading}"
+                    >
+                      ${question.id}
+                    </option>`,
                 )}
               </select>
               <div id="generation-results"></div>
@@ -83,8 +146,12 @@ export function AiGeneratePage({ resLocals }: { resLocals: Record<string, any> }
       </body>
       <script>
         function setPromptToExample() {
-          const prompt = document.getElementById('user-prompt-example').value;
-          document.getElementById('user-prompt-llm').value = prompt;
+          const options = document.getElementById('user-prompt-example').options;
+          const selection = options[options.selectedIndex].dataset;
+
+          document.getElementById('user-prompt-llm').value = selection.promptGeneral;
+          document.getElementById('user-prompt-llm-user-input').value = selection.promptUserInput;
+          document.getElementById('user-prompt-llm-grading').value = selection.promptGrading;
         }
       </script>
     </html>