+++ apps/prairielearn/src/pages/instructorQuestionSettings/instructorQuestionSettings.ts
@@ -14,7 +14,6 @@ import { generateSignedToken } from '@prairielearn/signed-token';
 import { b64EncodeUnicode } from '../../lib/base64-util.js';
 import { config } from '../../lib/config.js';
 import { copyQuestionBetweenCourses } from '../../lib/copy-question.js';
-import { IdSchema } from '../../lib/db-types.js';
 import {
   FileModifyEditor,
   QuestionRenameEditor,
@@ -30,6 +29,8 @@ import { startTestQuestion } from '../../lib/question-testing.js';
 import { encodePath } from '../../lib/uri-util.js';
 import { getCanonicalHost } from '../../lib/url.js';
 import { selectCoursesWithEditAccess } from '../../models/course.js';
+import { selectQuestionByUuid } from '../../models/question.js';
+import { selectTopicsByCourseId } from '../../models/topics.js';
 
 import {
   InstructorQuestionSettings,
@@ -62,6 +63,7 @@ router.post(
         res.locals.question,
         res.locals.course_instance,
         res.locals.course,
+        res.locals.user.user_id,
         res.locals.authn_user.user_id,
       );
       res.redirect(res.locals.urlPrefix + '/jobSequence/' + jobSequenceId);
@@ -78,6 +80,7 @@ router.post(
           res.locals.question,
           res.locals.course_instance,
           res.locals.course,
+          res.locals.user.user_id,
           res.locals.authn_user.user_id,
         );
         res.redirect(res.locals.urlPrefix + '/jobSequence/' + jobSequenceId);
@@ -112,6 +115,7 @@ router.post(
 
       const origHash = req.body.orig_hash;
       questionInfo.title = req.body.title;
+      questionInfo.topic = req.body.topic;
 
       const formattedJson = await formatJsonWithPrettier(JSON.stringify(questionInfo));
 
@@ -178,16 +182,17 @@ router.post(
         } catch {
           return res.redirect(res.locals.urlPrefix + '/edit_error/' + serverJob.jobSequenceId);
         }
-        const questionId = await sqldb.queryRow(
-          sql.select_question_id_from_uuid,
-          { uuid: editor.uuid, course_id: res.locals.course.id },
-          IdSchema,
-        );
+
+        const question = await selectQuestionByUuid({
+          course_id: res.locals.course.id,
+          uuid: editor.uuid,
+        });
+
         flash(
           'success',
           'Question copied successfully. You are now viewing your copy of the question.',
         );
-        res.redirect(res.locals.urlPrefix + '/question/' + questionId + '/settings');
+        res.redirect(res.locals.urlPrefix + '/question/' + question.id + '/settings');
       } else {
         await copyQuestionBetweenCourses(res, {
           fromCourse: res.locals.course,
@@ -206,39 +211,6 @@ router.post(
       } catch {
         res.redirect(res.locals.urlPrefix + '/edit_error/' + serverJob.jobSequenceId);
       }
-    } else if (req.body.__action === 'sharing_set_add') {
-      const questionSharingEnabled = await features.enabledFromLocals(
-        'question-sharing',
-        res.locals,
-      );
-      if (!questionSharingEnabled) {
-        throw new error.HttpStatusError(403, 'Access denied (feature not available)');
-      }
-      if (!res.locals.authz_data.has_course_permission_own) {
-        throw new error.HttpStatusError(403, 'Access denied (must be a course Owner)');
-      }
-      await sqldb.queryAsync(sql.sharing_set_add, {
-        course_id: res.locals.course.id,
-        question_id: res.locals.question.id,
-        unsafe_sharing_set_id: req.body.unsafe_sharing_set_id,
-      });
-      res.redirect(req.originalUrl);
-    } else if (req.body.__action === 'share_publicly') {
-      const questionSharingEnabled = await features.enabledFromLocals(
-        'question-sharing',
-        res.locals,
-      );
-      if (!questionSharingEnabled) {
-        throw new error.HttpStatusError(403, 'Access denied (feature not available)');
-      }
-      if (!res.locals.authz_data.has_course_permission_own) {
-        throw new error.HttpStatusError(403, 'Access denied (must be a course Owner)');
-      }
-      await sqldb.queryAsync(sql.update_question_shared_publicly, {
-        course_id: res.locals.course.id,
-        question_id: res.locals.question.id,
-      });
-      res.redirect(req.originalUrl);
     } else {
       throw new error.HttpStatusError(400, `unknown __action: ${req.body.__action}`);
     }
@@ -286,9 +258,12 @@ router.get(
       { question_id: res.locals.question.id },
       SelectedAssessmentsSchema,
     );
+
+    const courseTopics = await selectTopicsByCourseId(res.locals.course.id);
+
     const sharingEnabled = await features.enabledFromLocals('question-sharing', res.locals);
 
-    let sharingSetsIn, sharingSetsOther;
+    let sharingSetsIn;
     if (sharingEnabled) {
       const result = await sqldb.queryRows(
         sql.select_sharing_sets,
@@ -299,7 +274,6 @@ router.get(
         SharingSetRowSchema,
       );
       sharingSetsIn = result.filter((row) => row.in_set);
-      sharingSetsOther = result.filter((row) => !row.in_set);
     }
     const editableCourses = await selectCoursesWithEditAccess({
       user_id: res.locals.user.user_id,
@@ -328,11 +302,11 @@ router.get(
         assessmentsWithQuestion,
         sharingEnabled,
         sharingSetsIn,
-        sharingSetsOther,
         editableCourses,
         infoPath,
         origHash,
         canEdit,
+        courseTopics,
       }),
     );
   }),